Make scrolled() callback better readable and more efficient
The scrolled() callback made changes to the navigation buttons multiple
times the while the HTML view was updated during scrolling.
The nested for() loop and if() condition makes it hard to understand.

Now the update of the navigation buttens happens only once when switching
to a new slide. The nested for() if() construct is removed, making
the scolled() callback more readable.